Add support for minifying Javascript
This replicates the work that is currentby being done for Javascript
in socialbase/socialblue. The output of the js files as
tested is identical to that of the current gulp workflow with the
exception that the letters chosen by the minifier are different.
The code is functionally equivalent however.

The next step in this project is to also make this work for module
folders instead of theme folders.