feat: split library from cli component
This commit changes the structure of the repository by using lerna to
create a monorepo. We use this structure to split the CLI component that
is used in projects from the library component. This allows us to write
re-usable components that may be used by other projects without
downloading the CLI.

We configure webpack to leave all our dependencies external because the
code that we create is meant to be run using node.js and so any
dependencies should be installed by npm or yarn as needed.