feat(core): allow themes to be loaded by name
In order to be able to load parent themes we have to be able to find the
themes by name using the current theme's path as a starting point.

This commit adds methods that allow you to specify a theme name and a
path in a Drupal installation so that the theme can be located and
loaded.